Refactor helpers to common module
